chore: rename item-temp -> item
